+++ lib/core/utils/node-lookup.js
