+++ test/core/utils/node-lookup.js
